remove featuregate and old scraper implementation
