+++ receiver/dockerstatsreceiver/receiver_test.go
@@ -192,7 +192,6 @@ func TestScrapeV2(t *testing.T) {
 			cfg.Endpoint = tc.mockDockerEngine.URL
 			cfg.EnvVarsToMetricLabels = map[string]string{"ENV_VAR": "env-var-metric-label"}
 			cfg.ContainerLabelsToMetricLabels = map[string]string{"container.label": "container-metric-label"}
-			cfg.ProvidePerCoreCPUMetrics = true
 			cfg.MetricsConfig = allMetricsEnabled
 
 			receiver := newReceiver(receivertest.NewNopCreateSettings(), cfg)
